move ComposerPlugin.php out of src/ (#49)
`bin/hh-autoload` tries to parse everything under `src/`, but this file is PHP, not Hack, so it can't be parsed. It was actually parseable until a few days ago, but that was just luck.

Also removing `ComposerImporterTest` which relies on that file being parseable. New major release will remove all ComposerImporter stuff, but we want to release at least one more unbroken minor release.