Remove all mutable Vectors
This is always slightly risky.
Non of these vectors are used in a mutable position.
Meaning, there are no uses of this Vectors
that rely on using the Vector by changing it.
They are passed into functions, but those are all HSL functions.
These don't modify the Vector, so this is safe.
I have not accounted for unsafe operations which are invisible to hh_client.