Remove // partial
The legacy files have not worked since 4.67.
Top level code is not a thing anymore.
These files generate errors when disable_modes is true.
Bumped the minimum hhvm version
to not break for users that are relying on these files.
If composer installs this version for you, the file
would have thrown a top-level code \Error anyway.